+++ env_canada/constants.py
@@ -1,1 +1,1 @@
-USER_AGENT = "env_canada/0.12.3"
+USER_AGENT = "env_canada/0.12.4"